Destroy bucket only when TERRAFORM_DESTROY & TF_STATE_BUCKET_DESTROY = true

+++ operations/deployment/terraform/bitops.after-deploy.d/delete-tf-state-bucket.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-if [[ $TF_STATE_BUCKET_DESTROY = true ]] ; then
+if [[ $TERRAFORM_DESTROY = true ]] && [[ $TF_STATE_BUCKET_DESTROY = true ]]; then
   echo "Destroying S3 buket --> $TF_STATE_BUCKET"
   aws s3 rb s3://$TF_STATE_BUCKET --force
 fi
